Refactored core package
