+++ core/build.gradle
@@ -4,7 +4,7 @@ plugins {
 }
 
 android {
-    namespace 'com.artemissoftware.core_ui'
+    namespace 'com.artemissoftware.core'
     compileSdk rootProject.androidconfig.compileSdk
 
     defaultConfig {
