+++ core/src/androidTest/java/com/artemissoftware/core/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.core_ui
+package com.artemissoftware.core
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
