+++ core/src/main/java/com/artemissoftware/core/presentation/composables/TaskyCircleIdentifier.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.core_ui.composables
+package com.artemissoftware.core.presentation.composables
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
@@ -13,9 +13,9 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.artemissoftware.core_ui.composables.text.TaskyText
-import com.artemissoftware.core_ui.theme.Black
-import com.artemissoftware.core_ui.theme.Green
+import com.artemissoftware.core.presentation.composables.text.TaskyText
+import com.artemissoftware.core.presentation.theme.Black
+import com.artemissoftware.core.presentation.theme.Green
 
 @Composable
 fun TaskyCircleIdentifier(
