+++ core/src/main/java/com/artemissoftware/core/presentation/composables/button/TaskyButton.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.core_ui.composables.button
+package com.artemissoftware.core.presentation.composables.button
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -11,9 +11,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.artemissoftware.core_ui.composables.text.TaskyText
-import com.artemissoftware.core_ui.theme.Black
-import com.artemissoftware.core_ui.theme.White
+import com.artemissoftware.core.presentation.composables.text.TaskyText
+import com.artemissoftware.core.presentation.theme.Black
+import com.artemissoftware.core.presentation.theme.White
 
 @Composable
 fun TaskyButton(
