+++ core/src/main/java/com/artemissoftware/core/presentation/composables/textfield/TaskyOutlinedTextField.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.core_ui.composables.textfield
+package com.artemissoftware.core.presentation.composables.textfield
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.ExperimentalFoundationApi
@@ -27,9 +27,10 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.artemissoftware.core_ui.R
-import com.artemissoftware.core_ui.composables.text.TaskyText
-import com.artemissoftware.core_ui.theme.*
+import com.artemissoftware.core.R
+import com.artemissoftware.core.presentation.composables.text.TaskyText
+import com.artemissoftware.core.presentation.theme.*
+import com.artemissoftware.core.theme.*
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
