+++ core/src/main/java/com/artemissoftware/core/presentation/composables/textfield/TaskyTextFieldValidationStateType.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.core_ui.composables.textfield
+package com.artemissoftware.core.presentation.composables.textfield
 
 enum class TaskyTextFieldValidationStateType {
