+++ core/src/main/java/com/artemissoftware/core/presentation/theme/TaskyColor.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.core_ui.theme
+package com.artemissoftware.core.presentation.theme
 
 import androidx.compose.ui.graphics.Color
