+++ core/src/main/java/com/artemissoftware/core/presentation/theme/TaskyStyle.kt
@@ -1,4 +1,4 @@
-package com.artemissoftware.core_ui.theme
+package com.artemissoftware.core.presentation.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
@@ -6,7 +6,7 @@ import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import com.artemissoftware.core_ui.R
+import com.artemissoftware.core.R
 
 private val InterFont = FontFamily(
     Font(R.font.inter_regular),
