+++ core/src/main/res/drawable/ic_add.xml
