+++ core/src/main/res/drawable/ic_check.xml
