+++ core/src/main/res/drawable/ic_notification.xml
