+++ core/src/main/res/drawable/ic_visibility.xml
