+++ core/src/main/res/drawable/ic_visibility_off.xml
