+++ core/src/main/res/font/inter_bold.ttf
