+++ core/src/main/res/font/inter_extrabold.ttf
