+++ core/src/main/res/font/inter_light.ttf
