+++ core/src/main/res/font/inter_medium.ttf
