+++ core/src/main/res/font/inter_regular.ttf
