+++ core/src/main/res/font/inter_semibold.ttf
