+++ core/src/main/res/font/inter_thin.ttf
