+++ settings.gradle
@@ -14,4 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Tasky"
 include ':app'
-include ':core-ui'
+include ':core'
